(Feat) add company search
